// convex/_generated/api.d.ts
/* eslint-disable */
/**
 * Generated `api` utility.
 *
 * THIS CODE IS AUTOMATICALLY GENERATED.
 *
 * To regenerate, run `npx convex dev`.
 * @module
 */

import type * as conversations from "../conversations.js";
import type * as credits from "../credits.js";
import type * as files from "../files.js";
import type * as http from "../http.js";
import type * as imageGeneration from "../imageGeneration.js";
import type * as imageJobs from "../imageJobs.js";
import type * as llm from "../llm.js";
import type * as messages from "../messages.js";
import type * as models from "../models.js";
import type * as queries from "../queries.js";
import type * as seed from "../seed.js";
import type * as users from "../users.js";

import type {
  ApiFromModules,
  FilterApi,
  FunctionReference,
} from "convex/server";

declare const fullApi: ApiFromModules<{
  conversations: typeof conversations;
  credits: typeof credits;
  files: typeof files;
  http: typeof http;
  imageGeneration: typeof imageGeneration;
  imageJobs: typeof imageJobs;
  llm: typeof llm;
  messages: typeof messages;
  models: typeof models;
  queries: typeof queries;
  seed: typeof seed;
  users: typeof users;
}>;

/**
 * A utility for referencing Convex functions in your app's public API.
 *
 * Usage:
 * ```js
 * const myFunctionReference = api.myModule.myFunction;
 * ```
 */
export declare const api: FilterApi<
  typeof fullApi,
  FunctionReference<any, "public">
>;

/**
 * A utility for referencing Convex functions in your app's internal API.
 *
 * Usage:
 * ```js
 * const myFunctionReference = internal.myModule.myFunction;
 * ```
 */
export declare const internal: FilterApi<
  typeof fullApi,
  FunctionReference<any, "internal">
>;

export declare const components: {};